+++ dev/Ultima/World/Views/MiniMapBlock.cs
@@ -23,8 +23,8 @@ namespace UltimaXNA.Ultima.World.Views
             Colors = new uint[64];
 
             // get data from the tile Matrix
-            byte[] groundData = tileData.GetLandBlock((int)x, (int)y);
-            byte[] staticsData = tileData.GetStaticBlock((int)x, (int)y);
+            byte[] groundData = tileData.GetLandBlock(x, y);
+            byte[] staticsData = tileData.GetStaticBlock(x, y);
 
             // get the ground colors
             int groundDataIndex = 0;
